Move payload decoding logic out of server_log parser
